% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_nondiffuse.R
\name{adaptive_binarization}
\alias{adaptive_binarization}
\alias{adaptive_binarization,CanopyPhoto-method}
\title{Adaptive binarization}
\usage{
adaptive_binarization(cp, m = NULL, thr4only_catching_plants = 0.2)

\S4method{adaptive_binarization}{CanopyPhoto}(cp, m = NULL, thr4only_catching_plants = 0.2)
}
\arguments{
\item{cp}{\code{\linkS4class{CanopyPhoto}}. todo}

\item{m}{\linkS4class{BinImage}. todo}

\item{thr4only_catching_plants}{numerical.}
}
\value{
numeric or \code{\linkS4class{Raster}}.
}
\description{
todo
}
\details{
todo
}
\seealso{
\code{\link{autoThr}}, \code{\link{normalize}}.
}
