% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAutilityFunctions.R
\docType{methods}
\name{addMasks}
\alias{addMasks}
\alias{addMasks,BinImage-method}
\title{todo}
\usage{
addMasks(x, ...)

\S4method{addMasks}{BinImage}(x, ...)
}
\arguments{
\item{x}{todo}

\item{...}{todo}
}
\value{
todo
}
\description{
todo
}
\examples{
#/inst/examples/ Example.R

}
