% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_classes.R
\docType{class}
\name{Angle-class}
\alias{Angle-class}
\title{An S4 class to represent angle values.}
\description{
An S4 class to represent angle values. It has a slot that sets
  whether it is in degrees or radians, so the methods and functions can adapt
  to it instead of users. Only allows positive values. Use the helper
  function \code{\link{asAngle}} to generate new objects of this class.
}
\section{Slots}{

\describe{
\item{\code{values}}{numeric.}

\item{\code{degrees}}{logical.}
}}

\examples{
showClass("Angle")

}
\seealso{
\code{\link{asAngle}}
}
