% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAloadPhotoFunctions.R
\docType{methods}
\name{aperture}
\alias{aperture}
\alias{aperture<-}
\alias{aperture,CanopyPhoto-method}
\alias{aperture<-,CanopyPhoto,numeric-method}
\title{Set or get the aperture slot.}
\usage{
aperture(x)

\S4method{aperture}{CanopyPhoto}(x)

\S4method{aperture}{CanopyPhoto,numeric}(x) <- value
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{value}{Numeric. Aperture used to take \code{x}.}
}
\value{
numeric.
}
\description{
Set or get the the aperture used to take the photograph represented by a \code{\linkS4class{CanopyPhoto}}.
}
\examples{
x <- loadPhoto()
aperture(x)
aperture(x) <- 2.8
aperture(x)
}
\seealso{
\code{\link{loadPhoto}}.
}
