% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAutilityFunctions.R
\docType{methods}
\name{asAngle}
\alias{asAngle}
\alias{asAngle,numeric-method}
\title{Convert numeric to an Angle.}
\usage{
asAngle(x, degrees = TRUE)

\S4method{asAngle}{numeric}(x, degrees = TRUE)
}
\arguments{
\item{x}{numeric}

\item{degrees}{logical. By default is set as TRUE, meaning x is in degrees.
If FALSE x should be in radians.}
}
\value{
\code{\linkS4class{Angle}}.
}
\description{
Helper function for converting a numeric vector to an
  \code{\linkS4class{Angle}}. It is similar to coercing. For example,
  \code{as(45, "Angle")} return the same object than \code{asAngle(45)}.
}
\examples{
asAngle(90)
asAngle(c(1:90))

asAngle(pi) # this is incorrect
asAngle(pi, FALSE)

\dontrun{
  asAngle(370)
}
}
\seealso{
\code{\link{switchUnit}}
}
