% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_MBLT.R
\name{asynchronic_sky}
\alias{asynchronic_sky}
\alias{asynchronic_sky,data.frame-method}
\title{asynchronic sky}
\usage{
asynchronic_sky(block_info, path2sky_chunks, path4output, m, z, a)

\S4method{asynchronic_sky}{data.frame}(block_info, path2sky_chunks, path4output, m, z, a)
}
\arguments{
\item{block_info}{data.frame}

\item{path2sky_chunks}{character}

\item{path4output}{character}

\item{m}{BinImage}

\item{z}{ZenithImage}

\item{a}{AzimuthImage}
}
\description{
asynchronic sky.
}
\examples{
#TODO
}
