% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_classes.R
\docType{class}
\name{AzimuthImage-class}
\alias{AzimuthImage-class}
\title{An S4 class to represent azimuth angles as an image.}
\description{
An S4 class to represent azimuth angles as an image. Use the
  helper function \code{\link{makeAimage}} to generate new objects of this
  class.
}
\section{Slots}{

\describe{
\item{\code{...}}{Inherited from \code{\linkS4class{ZenithImage}}.}
}}

\examples{
showClass("AzimuthImage")

}
\seealso{
\code{\link{makeAimage}}
}
