% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbres.R
\docType{methods}
\name{bgInterpol}
\alias{bgInterpol}
\alias{bgInterpol,CanopyPhoto-method}
\title{Background interpolation}
\usage{
bgInterpol(x, bin, filling, prob = 0.95, erode = TRUE)

\S4method{bgInterpol}{CanopyPhoto}(x, bin, filling, prob = 0.95,
  erode = TRUE)
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{bin}{\code{\linkS4class{BinImage}}.}

\item{filling}{\code{\linkS4class{RasterLayer}} or \code{\linkS4class{RasterStack}}}

\item{prob}{numeric. Probability for \code[stats]{\link{quantile}}.}

\item{erode}{logical.}
}
\value{
\code{\linkS4class{BinImage}}
}
\description{
todo
}
\examples{
#todo
}
