% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_MBLT.R
\docType{methods}
\name{bgInterpol2}
\alias{bgInterpol2}
\alias{bgInterpol2,CanopyPhoto-method}
\title{Background interpolation}
\usage{
bgInterpol2(x, bin, mask, filling, prob = 0.95, erode = FALSE,
  fact = 5, np = 6)

\S4method{bgInterpol2}{CanopyPhoto}(x, bin, mask, filling, prob = 0.95,
  erode = FALSE, fact = 5, np = 6)
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{bin}{\code{\linkS4class{BinImage}}.}

\item{filling}{\code{\linkS4class{RasterLayer}} or \code{\linkS4class{RasterStack}}.}

\item{erode}{logical.}

\item{fact}{integer. Aggregation factor expressed as number of cells in each direction (horizontally and vertically). Or two integers (horizontal and vertical aggregation factor) or three integers (when also aggregating over layers). See Details}

\item{np}{
degree of polynomial surface
}
}
\value{
\code{\linkS4class{BinImage}}
}
\description{
todo
}
\examples{
#todo
}
