% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_MBLT.R
\name{bgModel}
\alias{bgModel}
\title{Background modelling}
\usage{
bgModel(x, z, a, bin, filling = NULL, prob = 0.95, erode = FALSE,
  ZA = TRUE, parallel = TRUE, freeThreads = 2)
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{z}{\code{\linkS4class{ZenithImage}}.}

\item{a}{\code{\linkS4class{AzimuthImage}}.}

\item{bin}{\code{\linkS4class{BinImage}}.}

\item{filling}{\code{\linkS4class{RasterLayer}}.}

\item{prob}{numeric. Probability for \code{\link[stats]{quantile}}.}

\item{erode}{logical.}

\item{ZA}{logical.}

\item{parallel}{logical.}

\item{freeThreads}{logical.}
}
\value{
\code{\linkS4class{BinImage}}
}
\description{
Backgroung modelling.
}
\examples{
#todo
}
