% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClasses.R
\docType{class}
\name{BinImage-class}
\alias{BinImage-class}
\title{An S4 class to store binarized images.}
\description{
An S4 class to store binarized images.
}
\section{Slots}{

\describe{
\item{\code{threshold}}{One-length numeric.}

\item{\code{originalData}}{One-length character.}

\item{\code{processedLayer}}{One-length integer    .}

\item{\code{...}}{Inherited from \code{\linkS4class{RasterLayer}}.}
}}

\examples{
showClass("BinImage")

}
\seealso{
\code{\link{autoThr}}, \code{\link{presetThr}},
  \code{\link{doMask}}, \code{\link{doOBIA}}
}
