% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAgeometryFunctions.R
\docType{methods}
\name{calcDiameter}
\alias{calcDiameter}
\alias{calcDiameter,CanopyPhoto,LensPolyCoef-method}
\title{Calculate diameter corresponding to a circular fisheye photography.}
\usage{
calcDiameter(x, y)

\S4method{calcDiameter}{CanopyPhoto,LensPolyCoef}(x, y)
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{y}{\code{\linkS4class{LensPolyCoef}}.}
}
\value{
numeric.
}
\description{
Calculate diameter corresponding to a circular fisheye photography.
}
\examples{
/inst/examples/calcDiameterExample.R
}
