% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAutilityFunctions.R
\docType{methods}
\name{calcExposure}
\alias{calcExposure}
\alias{calcExposure,numeric,numeric-method}
\title{Calculate exposure}
\usage{
calcExposure(ssDenominator, aperture)

\S4method{calcExposure}{numeric,numeric}(ssDenominator, aperture)
}
\arguments{
\item{ssDenominator}{numeric. Denominator of the shutter speed.}

\item{aperture}{numeric. The aperture or f number.}
}
\value{
numeric
}
\description{
Calculate the exposure for a given shutter speed an aperture.
}
\examples{
x <- loadPhoto()
calcExposure(ssDenominator(x), aperture(x))
}
\references{
Allen, E., & Triantaphillidou, S. (2010). The manual of
  photography. (E. Allen & S. Triantaphillidou, Eds.) (10th ed.). Amsterdam:
  Elsevier.
}
