% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_geometry.R
\docType{methods}
\name{calcOpticalCenter}
\alias{calcOpticalCenter}
\alias{calcOpticalCenter,data.frame-method}
\title{todo}
\usage{
calcOpticalCenter(x, y)

\S4method{calcOpticalCenter}{data.frame}(x)
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}
}
\value{
numeric. todo
}
\description{
todo
}
\examples{
/inst/examples/calcOpticalCenterExample.R
}
