% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_geometry.R
\name{calcR}
\alias{calcR}
\alias{calcR,Angle,LensPolyCoef-method}
\title{Calculate relative radius.}
\usage{
calcR(x, y)

\S4method{calcR}{Angle,LensPolyCoef}(x, y)
}
\arguments{
\item{x}{\code{\linkS4class{Angle}}.}

\item{y}{\code{\linkS4class{LensPolyCoef}}.}
}
\value{
numeric.
}
\description{
Given a zenith angle, it calculates the \strong{relative radius}
  (R) taken into account the lens distortion. For a definition of relative
  radius see details of \code{\link{lensPolyCoef}}.
}
\examples{
calcR(asAngle(10), lensPolyCoef())
calcR(asAngle(0), lensPolyCoef())
calcR(asAngle(90), lensPolyCoef())
calcR(asAngle(1:90), lensPolyCoef())
}
\seealso{
\code{\link{makeRimage}}, \code{\link{makeZimage}}
}
