% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_geometry.R
\name{calibrate_lens}
\alias{calibrate_lens}
\alias{calibrate_lens,data.frame-method}
\title{todo}
\usage{
calibrate_lens(csv)

\S4method{calibrate_lens}{data.frame}(csv)
}
\arguments{
\item{cvs}{todo}
}
\description{
todo
}
\examples{
#todo
}
