% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClasses.R
\docType{class}
\name{CanopyPhoto-class}
\alias{CanopyPhoto-class}
\title{An S4 class to store photographs of a vegetal canopy.}
\description{
An S4 class to store photographs of a vegetal canopy. Use the helper function \code{\link{loadPhoto}} to build new objects of this class from a file on disk.
}
\details{
\code{CanopyPhoto} can store photographs with any visual content but methods were designed to process real-color-photographs of the vegetal canopy.

\code{CanopyPhoto} was built on top of \code{RasterBrick} to take advantage of \code{raster package} functionality. However, \code{CanopyPhoto} require raster implementation but not georeferenced. That is why some inherited slots are meaningless for these class.
}
\section{Slots}{

\describe{
\item{\code{equipment}}{character.}

\item{\code{fisheye}}{\code{\linkS4class{FishEye}}.}

\item{\code{datetime}}{character.}

\item{\code{geocode}}{\code{\link[sp]{SpatialPoints}}.}

\item{\code{bearing}}{\code{\linkS4class{Angle}}.}

\item{\code{elevation}}{\code{\linkS4class{Angle}}.}

\item{\code{...}}{Inhereted from \code{\linkS4class{RasterBrick}}.}
}}
\examples{
showClass("CanopyPhoto")

}
\seealso{
\code{\link{loadPhoto}}
}

