% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAperPixelProcessingFunctions.R
\docType{methods}
\name{colorfulIndex}
\alias{colorfulIndex}
\alias{colorfulIndex,CanopyPhoto-method}
\title{Quantify the colorful of an image.}
\usage{
colorfulIndex(x, mask = NULL, plot = FALSE, returnRaster = FALSE)

\S4method{colorfulIndex}{CanopyPhoto}(x, mask = NULL, plot = FALSE,
  returnRaster = FALSE)
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}. Is the image that you want to know how colorful is.}

\item{mask}{\linkS4class{BinImage}. The default of NULL means that all the pixels will be taking into account in the computations. If you provide a \code{BinImage}, it must have the same extent and resolution of \code{x}. All pixels from the image covered by pixels of the \code{mask} with value \code{1} will be taking into account in the computations.}

\item{plot}{logical. The default is \code{FALSE}, if it is \code{TRUE} a plot will be printed. See Details.}

\item{returnRaster}{logical. The default is \code{FALSE}, if it is \code{TRUE} the raster that you can see when \code{plot} is \code{TRUE} is returned. See Details.}
}
\value{
Numeric or \code{\linkS4class{Raster}}.
}
\description{
Quantify the colorful of a sRGB image using a bidimensional space form by green/red and blue/yellow axes of the CIE \emph{L*a*b*} space, in other words, the lightness is excluded.  In that bidimensional space, this algorithm creates a square surface between -100 and 100 of both axes and see how much area of this square is covered by the image pixels. The index is the percentage of that covered area.
}
\details{
When you request a plot with argument \code{plot = TRUE}, the colorful index is returned but also a plot is send to the active graphics Device. The plot shows the color of the image represented in a bidimensional space former by axis \emph{a} and \emph{b} of the CIE \emph{L*a*b* space}.
}
\examples{
x <- loadPhoto()
x <- normalize(x, 0, 255)
z <- makeZimage(ncol(x), lensPolyCoef(c(0.6427, 0.0346, -0.024491)))
m <- doMask(z, alim = asAngle(c(0, 10)), zlim = asAngle(c(30, 40)))
m <- doMask(z)
colorfulIndex(x, m, TRUE)
r <- colorfulIndex(x, m, TRUE, TRUE)
plotRGB(r)
}
\seealso{
\code{link{normalize}}, \code{link{outOfDR}}.
}

