% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_read_cp.R
\docType{methods}
\name{elevation}
\alias{elevation}
\alias{elevation<-}
\alias{elevation,CanopyPhoto-method}
\alias{elevation<-,CanopyPhoto,Angle-method}
\title{Set or get the elevation slot.}
\usage{
elevation(x)

\S4method{elevation}{CanopyPhoto}(x)

\S4method{elevation}{CanopyPhoto,Angle}(x) <- value
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{value}{\code{\linkS4class{Angle}} that represents the elevation of
\code{x}.}
}
\value{
\code{\linkS4class{Angle}}.
}
\description{
Set or get the \code{\linkS4class{Angle}} that represents the
  elevation of a \code{\linkS4class{CanopyPhoto}}.
}
\details{
The elevation is the complement of the line-of-sight zenith angle.
  For zenith angle definition, see the details of \code{\link{lensPolyCoef}}.
}
\examples{
x <- loadPhoto()
elevation(x)
elevation(x) <- asAngle(55)
elevation(x)

\dontrun{
  elevation(x) <- asAngle(c(10,20))
}
}
\seealso{
\code{\link{loadPhoto}}.
}
