% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_read_cp.R
\name{equipment}
\alias{equipment}
\alias{equipment<-}
\alias{equipment,CanopyPhoto-method}
\alias{equipment<-,CanopyPhoto,character-method}
\title{Set or get the equipment slot.}
\usage{
equipment(x)

\S4method{equipment}{CanopyPhoto}(x)

\S4method{equipment}{CanopyPhoto,character}(x) <- value
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{value}{character that describe the equipment used to take \code{x}.}
}
\value{
\code{character}.
}
\description{
Set or get the length-one character string that describe the
  equipment used to take the photograph represented by a
  \code{\linkS4class{CanopyPhoto}}.
}
\examples{
x <- loadPhoto()
equipment(x)
equipment(x) <- "some string"
equipment(x)
equipment(x) <- c("some string 1", "some string 2")
equipment(x)
}
\seealso{
\code{\link{loadPhoto}}.
}
