% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAutilityFunctions.R
\docType{methods}
\name{expandFullframe}
\alias{expandFullframe}
\alias{expandFullframe,CanopyPhoto-method}
\title{Expand full frame hemispherical photographs}
\usage{
expandFullframe(x, zero = TRUE)

\S4method{expandFullframe}{CanopyPhoto}(x, zero = TRUE)
}
\arguments{
\item{x}{x \code{\linkS4class{CanopyPhoto}}.}

\item{zero}{logical. Default \code{TRUE} means that the added pixels will have \code{0}, \code{FALSE} means that will have \code{NA}.}
}
\value{
x \code{\linkS4class{CanopyPhoto}}.
}
\description{
Expand a full frame hemispherical photograph to get the equivalent of a circular hemispherical photograph in which the added pixels have \code{0} or \code{NA} in the tree channels.
}
\examples{
x <- loadPhoto()
z <- makeZimage(ncol(x), lensPolyCoef(c(0.9192, -0.1792, -0.000443)))
m <- doMask(z, zlim = asAngle(c(30, 60)))
plot(m)

m <- doMask(z, zlim = asAngle(c(30, 60)), alim = asAngle(c(0, 90)))
plot(m)

plot(x[m, drop = FALSE])

# Making an artificially fullframe hemiphoto by cropping a non-fullframe
# hemiphoto (just for making an example)
path <- system.file("external/UnFavAutoE3.jpg", package="caiman")
x <- loadPhoto(path, upperLeft = c(33, 120), width = 414, height = 240)
## declaring it as a fullframe
fisheye(x) <- newFishEye(TRUE, TRUE, TRUE)

m <- doMask(x)
plot(m)

m <- doMask(z, previousMask = m, zlim = asAngle(c(30, 60)),
                                                    alim = asAngle(c(0, 90)))
plot(m)

## Diagonals of the rectangular frame are the radius of the circle:
plot(doMask(makeRimage(ncol(m))) + m, legend = FALSE)
}
\seealso{
\code{\link{doMask}}.
}

