% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_read_cp.R
\name{exposureTime}
\alias{exposureTime}
\alias{exposureTime<-}
\alias{exposureTime,CanopyPhoto-method}
\alias{exposureTime<-,CanopyPhoto,numeric-method}
\title{Set or get the exposureTime slot.}
\usage{
exposureTime(x)

\S4method{exposureTime}{CanopyPhoto}(x)

\S4method{exposureTime}{CanopyPhoto,numeric}(x) <- value
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{value}{numeric. Denominator of the shutter speed used to take \code{x}.}
}
\value{
numeric.
}
\description{
Set or get the denominator of the shutter speed used to take the photograph represented by a \code{\linkS4class{CanopyPhoto}}.
}
\examples{
x <- loadPhoto()
exposureTime(x)
exposureTime(x) <- 1/125
exposureTime(x)
}
\seealso{
\code{\link{loadPhoto}}.
}
