% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBG.R
\docType{methods}
\name{extractBG}
\alias{extractBG}
\alias{extractBG,CanopyPhoto-method}
\alias{extractBG,data.frame-method}
\alias{extractBG,character-method}
\title{Assist the user to manually select pixels that only represent background}
\usage{
extractBG(x, ...)

\S4method{extractBG}{CanopyPhoto}(x, z = NULL, a = NULL)

\S4method{extractBG}{data.frame}(x, z = NULL, canopyPhoto, a = NULL)

\S4method{extractBG}{character}(x, id, z = NULL, mn, mx, a = NULL, ...)
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}, data.frame or character.}

\item{...}{Additional arguments as for \code{\link{loadPhoto}}:
\code{upperLeft}, \code{width} and \code{height}.}

\item{z}{\code{\linkS4class{ZenithImage}}. Optional.}

\item{a}{\code{\linkS4class{AzimuthImage}}. Optional.}

\item{canopyPhoto}{\code{\link{CanopyPhoto}}.}

\item{id}{Two columns data.frame. The first column is the name of the
photographs (extension included) and must be named \emph{photo}, the second
column is a factor and must be named \code{factor}.}

\item{mn}{numeric. Minimum expected value (see \code{\link{normalize}}).}

\item{mx}{numeric. Maximum expected value (see \code{\link{normalize}}).}
}
\value{
data.frame
}
\description{
Assist the user to manually select pixels that only represent background
}
\section{Methods (by class)}{
\itemize{
\item \code{CanopyPhoto}: Easy to use and useful if only one shot was
done per site.

\item \code{data.frame}: Easy to use and useful if you want to
extract the information of one photograph (\code{canopyPhoto}) with the
point that you manually selected on another (this makes sense if is the
same site and frame orientation). \code{x} is the data.frame produced using
\code{extractBG} and a given photograph.

\item \code{character}: s somewhat difficult to use but powerful. If
you taken photos in different sites and several pictures per site, is
probably your best option. \code{x} is the path to the folder that
contains all the photographs of the session.
}}

\examples{
# todo
}
