% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjfr.R
\name{extractBgb}
\alias{extractBgb}
\title{Extract background brightness}
\usage{
extractBgb(x, path)
}
\arguments{
\item{x}{data.frame. The return of a call to \code{\link{findOTRs}}.}

\item{path}{character. The path to the photographs of the canopy models. See
\code{\link{getReady4ECM}}.}
}
\value{
\code{\linkS4class{BinImage}}
}
\description{
Extract the background brightness of the photographs of the
  canopy models. See reference.
}
\examples{
#todo
}
\references{
todo
}
