% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSkyBrightness.R
\name{extractSkyBrightness}
\alias{extractSkyBrightness}
\title{todo}
\usage{
extractSkyBrightness(x, ...)
}
\arguments{
\item{x}{todo}

\item{...}{todo}
}
\value{
todo
}
\description{
todo
}
\examples{
todo
}
