% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractZenithBrightness.R
\name{extractZenithBrightness}
\alias{extractZenithBrightness}
\title{Extract sky brightness in zenith vecinity.}
\usage{
extractZenithBrightness(x, ...)
}
\arguments{
\item{x}{todo}

\item{...}{todo}

\item{z}{todo}

\item{upperLeft}{todo}

\item{width}{todo}

\item{samplingWindow}{todo}
}
\value{
data.frame
}
\description{
todo
}
\details{
todo
}
\examples{
"todo"
}
