% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbres.R
\docType{methods}
\name{findBias}
\alias{findBias}
\alias{findBias,CanopyPhoto-method}
\title{Find optimal bias}
\usage{
findBias(x, z, a, mapBg, thrFun, maxZ = asAngle(70))

\S4method{findBias}{CanopyPhoto}(x, z, a, mapBg, thrFun, maxZ = asAngle(70))
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{z}{\code{\linkS4class{ZenithImage}}.}

\item{a}{\code{\linkS4class{AzimuthImage}}.}

\item{mapBg}{\code{\linkS4class{RasterLayer}}.}

\item{thrFun}{function.}

\item{maxZ}{\code{\linkS4class{Angle}}.}
}
\value{
\code{\linkS4class{BinImage}}
}
\description{
Find optimal bias for get a gap on a given zenit angle.
}
\examples{
#todo
}
