% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_MBLT.R
\name{findOTRs}
\alias{findOTRs}
\alias{findOTRs,character-method}
\title{Find optimal threholding range}
\usage{
findOTRs(files, df, mask, percentE = NULL, absoluteE = 0.001, fun = mean, ...)

\S4method{findOTRs}{character}(files, df, mask, percentE = NULL, absoluteE = 0.001, fun = mean, ...)
}
\arguments{
\item{files}{character. The names of the canopy model photographs (extension
included), which were renamed by \code{\link{getReady4ECM}}. If your
working directory is the one that contains the photographs, use just the
file name, otherwise, use the full path to the file.}

\item{df}{data.frame.}

\item{mask}{\code{\linkS4class{BinImage}}. See \code{\link{doMask}}.}

\item{percentE}{numeric or NULL. Song et al. (2010) used 1 percent.}

\item{absoluteE}{numeric of NULL.}

\item{fun}{function. See \code{\link[raster]{calc}}}

\item{...}{aditionals arguments for \code{\link{loadPhoto}}.}
}
\value{
data.frame
}
\description{
Find optimal threholding range.
}
\examples{
#TODO
}
\references{
Song, G.-Z.M., Doley, D., Yates, D., Chao, K.-J., Hsieh, C.-F.,
  2014. Improving accuracy of canopy hemispherical photography by a constant
  threshold value derived from an unobscured overcast sky. Can. J. For. Res.
  44, 17-27. doi:10.1139/cjfr-2013-0082
}
