% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_classes.R
\docType{class}
\name{FishEye-class}
\alias{FishEye-class}
\title{An S4 class to represent the metadata related with fisheye photographs.}
\description{
An S4 class to represent the metadata related with fisheye photographs.
}
\section{Slots}{

\describe{
\item{\code{is}}{logical.}

\item{\code{up}}{logical.}

\item{\code{leveled}}{logical.}

\item{\code{fullframe}}{logical.}
}}

\examples{
showClass("FishEye")

}
\seealso{
\code{\link{newFishEye}}
}
