% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAloadPhotoFunctions.R
\docType{methods}
\name{fisheye}
\alias{fisheye}
\alias{fisheye,CanopyPhoto-method}
\alias{fisheye<-}
\alias{fisheye<-,CanopyPhoto,FishEye-method}
\title{Set or get fisheye.}
\usage{
fisheye(x)

\S4method{fisheye}{CanopyPhoto}(x)

\S4method{fisheye}{CanopyPhoto,FishEye}(x) <- value
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{value}{\code{\linkS4class{FishEye}} that contains the fisheye attributes of \code{x}.}
}
\value{
\code{\linkS4class{FishEye}}.
}
\description{
Set or get fisheye attributes for the photograph represented by a \code{\linkS4class{CanopyPhoto}}.
}
\examples{
x <- loadPhoto()
fisheye(x)
fisheye(x) <- newFishEye(TRUE, FALSE, FALSE)
fisheye(x)
}
\seealso{
\code{\link{loadPhoto}}.
}

