% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_geometry.R
\name{fisheye2pano}
\alias{fisheye2pano}
\alias{fisheye2pano,CanopyPhoto-method}
\alias{fisheye2pano,RasterLayer-method}
\title{todo}
\usage{
fisheye2pano(x, g, m = NULL, fill_na = TRUE, fun = mean)

\S4method{fisheye2pano}{CanopyPhoto}(x, g, m = NULL, fill_na = TRUE, fun = mean)

\S4method{fisheye2pano}{RasterLayer}(x, g, m = NULL, fill_na = TRUE, fun = mean)
}
\arguments{
\item{cp}{todo}
}
\description{
todo
}
\examples{
#todo
}
