% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheye2pano_dl.R
\name{fisheye2pano_dl}
\alias{fisheye2pano_dl}
\alias{fisheye2pano_dl,CanopyPhoto-method}
\alias{fisheye2pano_dl,RasterLayer-method}
\title{todo}
\usage{
fisheye2pano_dl(x, g, m = NULL, fill_na = TRUE, fun = mean)

\S4method{fisheye2pano_dl}{CanopyPhoto}(x, g, m = NULL, fill_na = TRUE, fun = mean)

\S4method{fisheye2pano_dl}{RasterLayer}(x, g, m = NULL, fill_na = TRUE, fun = mean)
}
\arguments{
\item{cp}{todo}
}
\description{
todo
}
\examples{
#todo
}
