% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_geometry.R
\docType{methods}
\name{fisheye2pano_hr}
\alias{fisheye2pano_hr}
\alias{fisheye2pano_hr,CanopyPhoto-method}
\title{todo}
\usage{
fisheye2pano_hr(hcp, z, a)

\S4method{fisheye2pano_hr}{CanopyPhoto}(hcp, z, a)
}
\arguments{
\item{cp}{todo}
}
\description{
todo
}
\examples{
#todo
}
