% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_read_cp.R
\docType{methods}
\name{fNumber}
\alias{fNumber}
\alias{fNumber<-}
\alias{fNumber,CanopyPhoto-method}
\alias{fNumber<-,CanopyPhoto,numeric-method}
\title{Set or get the fNumber slot.}
\usage{
fNumber(x)

\S4method{fNumber}{CanopyPhoto}(x)

\S4method{fNumber}{CanopyPhoto,numeric}(x) <- value
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{value}{Numeric. fNumber used to take \code{x}.}
}
\value{
numeric.
}
\description{
Set or get the the fNumber used to take the photograph represented by a \code{\linkS4class{CanopyPhoto}}.
}
\examples{
x <- loadPhoto()
aperture(x)
aperture(x) <- 2.8
aperture(x)
}
\seealso{
\code{\link{loadPhoto}}.
}
