% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_nondiffuse.R
\name{fuzzyLightness}
\alias{fuzzyLightness}
\alias{fuzzyLightness,numeric-method}
\title{Compute a weighted membership value to lightness.}
\usage{
fuzzyLightness(x, m, thr, fuzziness)

\S4method{fuzzyLightness}{numeric}(x, m, thr, fuzziness)
}
\arguments{
\item{x}{numeric. The lightness value.}

\item{m}{numeric lying between \code{0} and \code{1}, same length as
\code{x}. It is the scale parameter of the logistic membership function.}

\item{thr}{numeric of length \code{1}. Location parameter of the logistic
membership function.}

\item{fuzziness}{numeric of length \code{1}.}
}
\value{
\code{numeric.}
}
\description{
This algorithm uses a threshold value as the location parameter
  of a logistic membership function whose scale parameter depends on a
  variable. This dependence can be explained as follows: if the variable is
  equal to \code{1}, then the membership function is as a threshold function
  because the scale parameter is \code{0}; lowering the variable increases
  the scale parameter, thus blurring the threshold because it decreases the
  steepness of the curve.
}
\examples{
x <- raster(ncol = 100, nrow = 100)
extent(x) <- c(0, 100, 0, 100)
values(x) <- rep(1:ncol(x), nrow(x))
m <- t(x / getMax(x))
plot(x)
plot(m)

r <- x

values(r) <- fuzzyLightness(values(x), values(m), 45, 10)
plot(r, col = grey(1:100 / 100), ylab = "m * 100", xlab = "Lightness")

values(r) <- fuzzyLightness(values(x), values(m), 45, 50)
plot(r, col = grey(1:100 / 100), add = TRUE)

values(r) <- fuzzyLightness(values(x), values(m), 80, 20)
plot(r, col = grey(1:100 / 100), add = TRUE)

values(r) <- fuzzyLightness(values(x), values(m), 80, 1)
plot(r, col = grey(1:100 / 100), add = TRUE)
}
\references{
Diaz, G.M., Lencinas, J.D., 2015. Enhanced Gap Fraction
Extraction From Hemispherical Photography. IEEE Geosci. Remote Sens. Lett.
12, 1784-1789.
}
\seealso{
\code{\link[stats]{plogis}}.
}
