% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_read_cp.R
\docType{methods}
\name{geoLocation}
\alias{geoLocation}
\alias{geoLocation<-}
\alias{geoLocation,CanopyPhoto-method}
\alias{geoLocation<-,CanopyPhoto,SpatialPoints-method}
\title{Set or get the geoLocation slot.}
\usage{
geoLocation(x)

\S4method{geoLocation}{CanopyPhoto}(x)

\S4method{geoLocation}{CanopyPhoto,SpatialPoints}(x) <- value
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{value}{\code{\link[sp]{SpatialPoints}} that represents the location of
\code{x}.}
}
\value{
\code{\link[sp]{SpatialPoints}}.
}
\description{
Set or get the \code{\link[sp]{SpatialPoints}} that represent
  the shot location for the photograph represented by a
  \code{\linkS4class{CanopyPhoto}}.
}
\examples{
x <- loadPhoto()
geoLocation(x)
xy <- matrix(c(-57.95, -34.93), ncol = 2)
newPt <- SpatialPoints(xy, proj4string = CRS("+init=epsg:4326"))
geoLocation(x) <- newPt
geoLocation(x)

\dontrun{
  xy <- matrix(c(-57.95, -34.93, -57, -34), ncol = 2)
  newPt <- SpatialPoints(xy, proj4string = CRS("+init=epsg:4326"))
  geoLocation(x) <- newPt
}
}
\seealso{
\code{\link{loadPhoto}}.
}
