% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjfr.R
\docType{methods}
\name{getBgbfun}
\alias{getBgbfun}
\alias{getBgbfun,data.frame-method}
\title{Get a function to estimate the background brightness}
\usage{
getBgbfun(x, method = "BFGS", REVmn = NULL, REVmx = NULL)

\S4method{getBgbfun}{data.frame}(x, method = "BFGS", REVmn = NULL,
  REVmx = NULL)
}
\arguments{
\item{x}{data.frame. The return of a call to \code{\link{extractBgb}}.}

\item{method}{character. See \code{\link[bbmle]{mle2}}}

\item{REVmn}{numeric.}

\item{REVmx}{numeric.}
}
\value{
list
}
\description{
Wrapper function for \code{bbmle::mle2} to make easier the estimation
  of a function to estimate the background brightness.
}
\examples{
#TODO
}
\references{
todo
}
