% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMfunctions.R
\docType{methods}
\name{getECMfun}
\alias{getECMfun}
\alias{getECMfun,data.frame-method}
\title{Get the function for EC method}
\usage{
getECMfun(x, method = "BFGS")

\S4method{getECMfun}{data.frame}(x, method = "BFGS")
}
\arguments{
\item{x}{data.frame. The return of a call to \code{\link{findOTRs}}.}

\item{method}{character. See \code{\link[bbmle]{mle2}}}
}
\value{
list
}
\description{
Wrapper function for \code{bbmle::mle2} to make easier the estimation
  of the function for the EC method.
}
\examples{
#TODO
}
\references{
Song, G.-Z.M., Doley, D., Yates, D., Chao, K.-J., Hsieh, C.-F.,
  2014. Improving accuracy of canopy hemispherical photography by a constant
  threshold value derived from an unobscured overcast sky. Can. J. For. Res.
  44, 17–27. doi:10.1139/cjfr-2013-0082
}
