% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_object_based.R
\docType{methods}
\name{getFeatureImage}
\alias{getFeatureImage}
\alias{getFeatureImage,RasterLayer-method}
\title{Get an image that stores object features.}
\usage{
getFeatureImage(x, ...)

\S4method{getFeatureImage}{RasterLayer}(x, segmentation, fun)
}
\arguments{
\item{x}{\code{\linkS4class{RasterLayer}}. The image that you want to extract
feature from.}

\item{...}{Additional arguments (none implemented).}

\item{segmentation}{\linkS4class{PolarSegmentation}. The segmentation of
\code{x}.}

\item{fun}{\code{function} that take a vector as input and return a one-length
numeric or logical (e.g. mean).}
}
\value{
\code{\linkS4class{RasterLayer}}.
}
\description{
Given a single layer data, a segmentation and a function, this
  algorithm returns \code{RasterLayer} like segmentation but with its segment-IDs
  replaced with the segment-features, which are the return of processing all
  the pixels that belong to a segment with the aforementioned function.
}
\examples{
x <- loadPhoto()
z <- makeZimage(ncol(x), lensPolyCoef(c(0.9192, -0.1792, -0.000443)))
r <- makeRings(z, angleWidth = asAngle(30))
fi <- getFeatureImage(raster::subset(x, 3), r, sd)
plot(fi)
}
\seealso{
\code{\link{extractFeatures}}.
}
