% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAloadPhotoFunctions.R
\docType{methods}
\name{getHour}
\alias{getHour}
\alias{getHour,character-method}
\title{Get hour of a day from datetime character}
\usage{
getHour(x)

\S4method{getHour}{character}(x)
}
\arguments{
\item{x}{character.}
}
\value{
numeric.
}
\description{
Get hour of a day from datetime character.
}
\examples{
x <- loadPhoto()
x <- datetime(x)
x
getHour(x)

}
\seealso{
\code{\link{datetime}}.
}
