% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAutilityFunctions.R
\docType{methods}
\name{getMin}
\alias{getMax}
\alias{getMax,RasterAll-method}
\alias{getMin}
\alias{getMin,RasterAll-method}
\title{Get the extreme values.}
\usage{
getMin(x)

\S4method{getMin}{RasterAll}(x)

getMax(x)

\S4method{getMax}{RasterAll}(x)
}
\arguments{
\item{x}{\code{\linkS4class{Raster}}.}
}
\value{
A numeric vector.
}
\description{
Get the maximum/minimum value of a \code{\linkS4class{Raster}}. If it has multiple layers get the maximums/minimums values of each one.
}
\examples{
x <- loadPhoto()
getMax(x)
getMin(x)

x <- makeRimage(480)
getMax(x)
getMin(x)
}

