% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAutilityFunctions.R
\docType{methods}
\name{getNumberOfPixels}
\alias{getNumberOfPixels}
\alias{getNumberOfPixels,PolarSegmentation-method}
\title{todo}
\usage{
getNumberOfPixels(x, m, onlyMin = TRUE)

\S4method{getNumberOfPixels}{PolarSegmentation}(x, m, onlyMin = TRUE)
}
\arguments{
\item{x}{todo}

\item{m}{todo}

\item{onlyMin}{todo}
}
\value{
todo
}
\description{
todo
}
\examples{
#/inst/examples/ Example.R

}
