% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_MBLT.R
\docType{methods}
\name{getReady4ECM}
\alias{getReady4ECM}
\alias{getReady4ECM,character-method}
\title{getReady4ECM}
\usage{
getReady4ECM(files, keyPhotos, openPipePhotos, openness, path)

\S4method{getReady4ECM}{character}(files, keyPhotos, openPipePhotos,
  openness, path)
}
\arguments{
\item{files}{character. The names of the photographs (extension included) of
the canopy models. If your working directory is the one that contains the
photographs, use just the file name, otherwise, use the full path to
the file.}

\item{keyPhotos}{numeric vector. The numeric ID of the first photograph taken
to a canopy model with a given light.}

\item{openPipePhotos}{numeric vector.}

\item{openness}{numeric.}

\item{path}{character.}
}
\value{
data.frame
}
\description{
todo.
}
\examples{
#TODO
}
\references{
Song, G.-Z.M., Doley, D., Yates, D., Chao, K.-J., Hsieh, C.-F.,
  2014. Improving accuracy of canopy hemispherical photography by a constant
  threshold value derived from an unobscured overcast sky. Can. J. For. Res.
  44, 17–27. doi:10.1139/cjfr-2013-0082
}
