% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ_group_untested.R
\docType{methods}
\name{gf4cimes}
\alias{gf4cimes}
\alias{gf4cimes,BinImage-method}
\alias{gf4cimes,character-method}
\title{todo}
\usage{
gf4cimes(x, ...)

\S4method{gf4cimes}{BinImage}(x, z, a, m, angleWidth, file, txt = TRUE)

\S4method{gf4cimes}{character}(x, z, a, m, subfolder = "out", ...)
}
\arguments{
\item{x}{todo}

\item{...}{todo}

\item{z}{todo}

\item{a}{todo}

\item{m}{todo}

\item{angleWidth}{todo}

\item{file}{todo}

\item{txt}{todo}
}
\value{
todo
}
\description{
todo
}
\examples{
todo
}
