% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ_group_untested.R
\name{gs4cimes}
\alias{gs4cimes}
\alias{gs4cimes,BinImage-method}
\alias{gs4cimes,character-method}
\title{todo}
\usage{
gs4cimes(x, ...)

\S4method{gs4cimes}{BinImage}(x, z, a, m, file, forceEqualLength = FALSE, txt = TRUE)

\S4method{gs4cimes}{character}(x, z, a, ...)
}
\arguments{
\item{x}{todo}

\item{...}{todo}

\item{z}{todo}

\item{a}{todo}

\item{m}{todo}

\item{file}{todo}

\item{forceEqualLength}{todo}

\item{txt}{todo}
}
\value{
todo
}
\description{
todo
}
\examples{
#todo
}
