% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_per_pixel.R
\name{HLS2sRGB}
\alias{HLS2sRGB}
\alias{HLS2sRGB,matrix-method}
\alias{HLS2sRGB,numeric-method}
\alias{HLS2sRGB,CanopyPhoto-method}
\title{Convert sRGB to HLS.}
\usage{
HLS2sRGB(x, ...)

\S4method{HLS2sRGB}{matrix}(x)

\S4method{HLS2sRGB}{numeric}(x)

\S4method{HLS2sRGB}{CanopyPhoto}(x, ...)
}
\arguments{
\item{x}{numeric, matrix or \code{\linkS4class{CanopyPhoto}}. Values must
lying between \code{0} and \code{1}.}

\item{...}{Additional arguments as for \code{\link[raster]{writeRaster}}.}
}
\value{
Matrix, or \code{\linkS4class{CanopyPhoto}}.
}
\description{
Wrapper function of \code{\link[colorspace]{HLS}} and
  \code{\link[colorspace]{sRGB}} that convert colors from HLS to sRGB.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: Each row should represent a color. Argument \code{x} must have three columns.

\item \code{numeric}: Convert a single color from HLS to sRGB. Argument \code{x} must be of length three. The output is a matrix.

\item \code{CanopyPhoto}: The output is a \code{\linkS4class{CanopyPhoto}} with layer names Red, Green and Blue.
}}

\seealso{
\code{\link{normalize}}.

example /inst/examples/HLS2sRGBexample.R
}
