% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_read_cp.R
\name{isoSpeed}
\alias{isoSpeed}
\alias{isoSpeed<-}
\alias{isoSpeed,CanopyPhoto-method}
\alias{isoSpeed<-,CanopyPhoto,numeric-method}
\title{Set or get the isoSpeed slot.}
\usage{
isoSpeed(x)

\S4method{isoSpeed}{CanopyPhoto}(x)

\S4method{isoSpeed}{CanopyPhoto,numeric}(x) <- value
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{value}{numeric. ISO Speed used to take \code{x}.}
}
\value{
numeric.
}
\description{
Set or get the ISO speed used to take the photograph represented by a \code{\linkS4class{CanopyPhoto}}.
}
\examples{
x <- loadPhoto()
isoSpeed(x)
isoSpeed(x) <- 100
isoSpeed(x)
}
\seealso{
\code{\link{loadPhoto}}.
}
