% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_classes.R
\docType{class}
\name{LensPolyCoef-class}
\alias{LensPolyCoef-class}
\title{An S4 class to represent coefficients that models lens distortion.}
\description{
An S4 class to represent coefficients that models lens
  distortion of hemispherical photographs. Use the helper function
  \code{\link{lensPolyCoef}} to generate new objects of this class.
}
\section{Slots}{

\describe{
\item{\code{coef}}{numeric.}
}}

\examples{
showClass("LensPolyCoef")

}
\seealso{
\code{\link{lensPolyCoef}}
}
