% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_geometry.R
\docType{methods}
\name{makeAimage}
\alias{makeAimage}
\alias{makeAimage,ZenithImage-method}
\title{Generate an AzimuthImage.}
\usage{
makeAimage(x)

\S4method{makeAimage}{ZenithImage}(x)
}
\arguments{
\item{x}{\code{\linkS4class{ZenithImage}}.}
}
\value{
\code{\linkS4class{AzimuthImage}}.
}
\description{
Helper function to generate an
  \code{\linkS4class{AzimuthImage}}. Compute Azimuth angle for each pixels
  that are needed to represent a circular hemispherical image with \code{180}
  degrees field of view. North is oriented up and West is oriented left.
}
\details{
Azimuth angle is the angle between the North direction and the line
  of sight projected to the leveled plane. The angle increases in clockwise
  direction at the field but not in the hemispherical image because it is
  flipped in the East-West direction.
}
\examples{
z <- makeZimage(480, lensPolyCoef())
a <- makeAimage(z)
plot(a)
a
}
\seealso{
\code{\link{makeZimage}}.
}
