% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AApolarSegmentationAlgorithms.R
\docType{methods}
\name{makePolarSectors}
\alias{makePolarSectors}
\alias{makePolarSectors,AzimuthImage-method}
\title{Slice the polar space in sectors}
\usage{
makePolarSectors(x, angleWidth, angleMean = FALSE)

\S4method{makePolarSectors}{AzimuthImage}(x, angleWidth, angleMean = FALSE)
}
\arguments{
\item{x}{\code{\linkS4class{AzimuthImage}}.}

\item{angleWidth}{\code{\linkS4class{Angle}}. AngleWidth must divide 0-360 into a whole number of segments.}

\item{angleMean}{logical. If it is FALSE, all the pixels belong to a sector are labeled with an ID number. Otherwise, the angle mean of the sector are assigned to the pixels.}
}
\value{
\code{\linkS4class{PolarSegmentation}}.
}
\description{
Make sectors by slicing the azimuth angle from 0 to 360 in equals intervals.
}
\details{
The intervals are closed on the right and open on the left. The first sector never contain \code{0} because zenith point is always in between two pixels (\code{\link{makeZimage}}), that is way a pixel never will be in the exact Norte direction.
}
\examples{
z <- makeZimage(480, lensPolyCoef())
a <- makeAimage(z)

s <- makePolarSectors(a, angleWidth = asAngle(10))
s
plot(s)
s <- makePolarSectors(a, angleWidth = asAngle(10), angleMean = TRUE)
s
plot(s)

s <- makePolarSectors(a, angleWidth = asAngle(15))
s
plot(s)
}
\seealso{
\code{\link{makeAimage}}, \code{\link{makeRings}}, \code{\link{makePolarGrid}}.
}

