% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AApolarSegmentationAlgorithms.R
\docType{methods}
\name{makeRings}
\alias{makeRings}
\alias{makeRings,ZenithImage-method}
\title{Make rings.}
\usage{
makeRings(x, angleWidth, angleMean = FALSE)

\S4method{makeRings}{ZenithImage}(x, angleWidth, angleMean = FALSE)
}
\arguments{
\item{x}{\code{\linkS4class{ZenithImage}}.}

\item{angleWidth}{\code{\linkS4class{Angle}}. It must divide 0-90
into a whole number of segments.}

\item{angleMean}{logical. If \code{FALSE}, all the pixels that belong to a ring are
labeled with an ID number. Otherwise, the angle mean of the ring is
assigned to the pixels.}
}
\value{
\code{\linkS4class{PolarSegmentation}}.
}
\description{
Make rings by dividing the zenith angle from \code{0} to
  \code{90} in equals intervals.
}
\details{
The intervals are closed on the right and open on the left. The
  first ring never contains \code{0} because the zenith point is always in between
  two pixels (see \code{\link{makeZimage}}).
}
\examples{
z <- makeZimage(480, lensPolyCoef())

r <- makeRings(z, angleWidth = asAngle(30))
r
plot(r)
r <- makeRings(z, angleWidth = asAngle(30), angleMean = TRUE)
r
plot(r)

r <- makeRings(z, angleWidth = asAngle(10))
plot(r)
}
\seealso{
\code{\link{makeZimage}}, \code{\link{makePolarSectors}}, \code{\link{makePolarGrid}}.
}
