% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAutilityFunctions.R
\docType{methods}
\name{masking}
\alias{masking}
\alias{masking,RasterLayer-method}
\alias{masking,RasterStackBrick-method}
\title{todo}
\usage{
masking(x, m, col = c(1, 0, 0))

\S4method{masking}{RasterLayer}(x, m, col = c(1, 0, 0))

\S4method{masking}{RasterStackBrick}(x, m, col = c(1, 0, 0))
}
\arguments{
\item{x}{todo}

\item{m}{todo}

\item{col}{todo}
}
\value{
todo
}
\description{
todo
}
\examples{
#/inst/examples/ Example.R

}
