% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_MBLT.R
\docType{methods}
\name{MBLT1}
\alias{MBLT1}
\alias{MBLT1,CanopyPhoto-method}
\title{Local thresholding through sky DN estimation}
\usage{
MBLT1(ncp, z, a, m, w = 0.5, filling_source = NULL, sky = FALSE,
  linear_stretching = FALSE, thr_fun = preset_thr_fun,
  channel = "Blue")

\S4method{MBLT1}{CanopyPhoto}(ncp, z, a, m, w = 0.5,
  filling_source = NULL, sky = FALSE, linear_stretching = FALSE,
  thr_fun = preset_thr_fun, channel = "Blue")
}
\arguments{
\item{ncp}{\code{\linkS4class{CanopyPhoto}}.}

\item{z}{todo}

\item{a}{todo}

\item{m}{todo}

\item{w}{todo}

\item{filling_source}{todo}

\item{thr_fun}{todo}

\item{channel}{todo}
}
\value{
\code{\linkS4class{BinImage}} or \code{\linkS4class{RasterLayer}}
}
\description{
todo
}
\examples{
#todo
}
