% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_MBLT.R
\name{MBLT2}
\alias{MBLT2}
\alias{MBLT2,CanopyPhoto-method}
\title{todo}
\usage{
MBLT2(
  ncp,
  m,
  bin_plant,
  bin_sky,
  fill = NULL,
  sample_size = 30,
  size_range = NULL,
  size_step = NULL,
  channel = "Blue",
  parallel = TRUE,
  free_threads = 2
)

\S4method{MBLT2}{CanopyPhoto}(
  ncp,
  m,
  bin_plant,
  bin_sky,
  fill = NULL,
  sample_size = 30,
  size_range = NULL,
  size_step = NULL,
  channel = "Blue",
  parallel = TRUE,
  free_threads = 2
)
}
\arguments{
\item{ncp}{\code{\linkS4class{CanopyPhoto}}.}

\item{m}{todo}

\item{bin_plant}{todo}

\item{bin_sky}{todo}

\item{fill}{todo}

\item{sample_size}{must be even}

\item{size_range}{todo}

\item{channel}{todo}

\item{parallel}{todo}

\item{free_threads}{todo}
}
\value{
\code{\linkS4class{BinImage}}
}
\description{
todo
}
\examples{
#todo
}
