% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelBasedThresholding.R
\name{modelBasedThresholding}
\alias{modelBasedThresholding}
\title{todo}
\usage{
modelBasedThresholding(x, ...)
}
\arguments{
\item{x}{todo}

\item{...}{todo}

\item{z}{todo}

\item{path2photos}{todo}

\item{path4write}{todo}

\item{fun}{todo}

\item{model}{todo}

\item{mn}{todo}

\item{mx}{todo}
}
\value{
todo
}
\description{
todo
}
\examples{
#todo
}
