% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAloadPhotoFunctions.R
\docType{methods}
\name{newFishEye}
\alias{newFishEye}
\alias{newFishEye,logical,logical,logical-method}
\alias{newFishEye,missing,missing,missing-method}
\title{Generate a new FishEye}
\usage{
newFishEye(up, leveled, fullframe)

\S4method{newFishEye}{logical,logical,logical}(up, leveled, fullframe)

\S4method{newFishEye}{missing,missing,missing}(up, leveled, fullframe)
}
\arguments{
\item{up}{missing or logical. Set \code{TRUE} if the photograph was taken looking up.}

\item{leveled}{missing or logical. Set \code{TRUE} if the camera was leveled and therefore its optical axis was aligned with a perfect vertical.}

\item{fullframe}{missing or logical. Set \code{TRUE} if all the pixels of the photograph have data. For circular fisheye, with hundreds of no-data pixels, set \code{FALSE}.}
}
\value{
\code{\linkS4class{FishEye}}.
}
\description{
Helper function to generate a \code{\linkS4class{FishEye}} needed to fill the slot \code{filsheye} of \code{\linkS4class{CanopyPhoto}}.
}
\details{
A call to \code{newFishEye()} return the prototype for \code{\linkS4class{FishEye}}.
}
\examples{
newFishEye()
newFishEye(TRUE, TRUE, FALSE)
}
\seealso{
\code{\link{loadPhoto}}.
}

