% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ_group_untested.R
\docType{methods}
\name{nh2005}
\alias{nh2005}
\alias{nh2005,CanopyPhoto-method}
\alias{nh2005,RasterLayer-method}
\title{Experimental.}
\usage{
nh2005(x, parallel = TRUE, freeThreads = 1, ...)

\S4method{nh2005}{CanopyPhoto}(x, subset = 3)

\S4method{nh2005}{RasterLayer}(x, parallel, freeThreads)
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}} or
\code{\linkS4class{RasterLayer}}.}

\item{...}{Additional arguments (none implemented).}

\item{subset}{numeric or character. It indicates the channel to process
(represented as integer or by their name).}
}
\value{
\code{\linkS4class{BinImage}}.
}
\description{
Experimental.
}
\section{Methods (by class)}{
\itemize{
\item \code{CanopyPhoto}: The argument \code{subset} will be passed to
\code{\link[raster]{subset}} for selecting which channel of the
\code{\linkS4class{CanopyPhoto}} will be processed. It computes the
threshold of the selected layer and return a \code{\linkS4class{BinImage}}.

\item \code{RasterLayer}: Compute the threshold and return a
\code{\linkS4class{BinImage}}.
}}

\seealso{
\code{\link{presetThr}}.
}
