% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_per_pixel.R
\docType{methods}
\name{outOfDR}
\alias{outOfDR}
\alias{outOfDR,CanopyPhoto-method}
\title{Get the percentages of pixels that are out of the dynamic range.}
\usage{
outOfDR(x, channel = NULL, mask = NULL, returnImages = FALSE)

\S4method{outOfDR}{CanopyPhoto}(x, channel = NULL, mask = NULL,
  returnImages = FALSE)
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{channel}{One-length character, "Red", "Green" or "Blue".}

\item{mask}{\linkS4class{BinImage}. Default value \code{NULL} means that all
the pixels will be taking into account in the computations. If you provide
a \linkS4class{BinImage}, it must have the same extent and resolution of
\code{x}. All pixels from the image covered by pixels of the mask with
value \code{1} will be taking into account in the computations.}

\item{returnImages}{logical. Default is \code{FALSE}, see Value.}
}
\value{
By default, it returns a vector of length \code{3}. If
  \code{returnImages = TRUE}, then it returns a
  \code{\linkS4class{RasterStack}} that has two binary layers in which \code{1}
  means under/overexposure.

  todo
}
\description{
Get the percentages of pixels that are out of the dynamic range,
  i.e., the under and overexposure pixels of an image.
}
\details{
This algorithm classifies the pixels of x that start and end the
  dynamic range. These pixels are called under or overexposed because is
  highly probable that they are out of the dynamic range and not just in the
  limits.

  todo
}
\examples{
x <- loadPhoto()
x <- normalize(x, 0, 255)
z <- makeZimage(ncol(x), lensPolyCoef())
m <- doMask(z)

outOfDR(x, mask = m)
plot(outOfDR(x, mask = m,  returnImages = TRUE))

outOfDR(x, "Blue", mask = m)
plot(outOfDR(x, "Blue", mask = m,  returnImages = TRUE))
}
\seealso{
\code{\link{normalize}}, \code{\link{colorfulness}}.
}
