% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_classes.R
\docType{class}
\name{PolarSegmentation-class}
\alias{PolarSegmentation-class}
\title{An S4 class to store polar segmentations.}
\description{
An S4 class to store polar segmentations.
}
\section{Slots}{

\describe{
\item{\code{angleWidth}}{\code{\linkS4class{Angle}}}

\item{\code{scaleParameter}}{numeric}

\item{\code{...}}{Inherited from \code{\linkS4class{ZenithImage}}.}
}}

\examples{
showClass("PolarSegmentation")

}
\seealso{
\code{\link{makeRings}}, \code{\link{makePolarSectors}},
  \code{\link{makePolarGrid}}
}
