% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_MBLT.R
\docType{methods}
\name{preset_thr_fun}
\alias{preset_thr_fun}
\alias{preset_thr_fun,numeric-method}
\title{todo}
\usage{
preset_thr_fun(DN, w = 1, type = "Generic")

\S4method{preset_thr_fun}{numeric}(DN, w = 1, type = "Generic")
}
\arguments{
\item{DN}{numeric}

\item{w}{numeric}

\item{type}{character}
}
\description{
todo
}
\examples{
#todo
}
