% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_per_pixel.R
\name{presetThr}
\alias{presetThr}
\alias{presetThr,RasterLayer-method}
\title{Binarize images with a preset threshold.}
\usage{
presetThr(x, ...)

\S4method{presetThr}{RasterLayer}(x, thr)
}
\arguments{
\item{x}{\code{\linkS4class{RasterLayer}}.}

\item{...}{Additional arguments (none implemented)}

\item{thr}{numeric. Threshold value. It should be between minimum and maximum layer value.}
}
\value{
\code{\linkS4class{BinImage}}.
}
\description{
Given a single layer image and a global threshold value, it
  returns a binarized image. All pixels above the threshold are replaced with
  \code{1} and the rest with \code{0}.
}
\examples{
x <- loadPhoto()
x <- raster::subset(x, 3)
bin <- presetThr(x, 125)
bin
plot(bin)
}
\seealso{
\code{\link{autoThr}}.
}
