% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_MBLT.R
\docType{methods}
\name{process_multi_point}
\alias{process_multi_point}
\alias{process_multi_point,data.frame-method}
\title{process multi point}
\usage{
process_multi_point(block_info, path2data, path4output,
  datatype = "INT2U")

\S4method{process_multi_point}{data.frame}(block_info, path2data,
  path4output, datatype = "INT2U")
}
\arguments{
\item{block_info}{data.frame}

\item{path2data}{character}

\item{path4output}{character}

\item{datatype}{character}
}
\description{
process multi point.
}
\examples{
#TODO
}
