% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_utilities.R
\docType{methods}
\name{read_bin}
\alias{read_bin}
\alias{read_bin,character-method}
\title{todo}
\usage{
read_bin(filename)

\S4method{read_bin}{character}(filename)
}
\arguments{
\item{filename}{todo}
}
\value{
todo
}
\description{
todo
}
\examples{
#/inst/examples/ Example.R

}
