% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_object_based.R
\name{regional_thr}
\alias{regional_thr}
\alias{regional_thr,RasterLayer-method}
\title{Regional Thresholding}
\usage{
regional_thr(r, segmentation, method = "autoThr")

\S4method{regional_thr}{RasterLayer}(r, segmentation, method = "autoThr")
}
\arguments{
\item{r}{\code{\linkS4class{RasterLayer}}.}

\item{segmentation}{\linkS4class{PolarSegmentation}. The segmentation of
\code{x}.}

\item{method}{\code{character}.}

\item{...}{Additional arguments (none implemented).}
}
\description{
todo
}
