% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_classes.R
\docType{class}
\name{RelativeRadiusImage-class}
\alias{RelativeRadiusImage-class}
\title{An S4 class to represent Relative Radius as an image.}
\description{
An S4 class to represent Relative Radius as an image. Use the
  helper function \code{\link{makeRimage}} to generate a new
  \code{\linkS4class{RelativeRadiusImage}}.
}
\details{
\code{raster} package offer a wide range of functionalities to
  manipulate raster files efficiently. It uses georeferenced raster because
  was designed with a field-based conception of the geographical reality
  (Galton, 2011). To manipulate such data, \code{raster} package defines "S4
  classes" grouped in the \code{\linkS4class{Raster}} family that has a complex hierarchical
  structure of inheritance. These classes have slots that help to precisely
  locate each cell over the earth because they were designed to represent
  georeferenced raster. \code{RelativeRadiusImage} was built on top of \code{RasterLayer}
  to take advantage of \code{raster} package functionality. However,
  \code{RelativeRadiusImage} requires raster implementation but not georeference.
  That is why some inherited slots are meaningless for these class.
}
\section{Slots}{

\describe{
\item{\code{diameter}}{integer.}

\item{\code{...}}{Inherited from \code{\linkS4class{RasterLayer}}.}
}}

\examples{
showClass("RelativeRadiusImage")

}
\references{
Galton, A., 2001. A Formal Theory of Objects and Fields, in:
  COSIT. pp. 458-473.
}
\seealso{
\code{\link{makeRimage}}
}
