% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_geometry.R
\docType{methods}
\name{rotAzim}
\alias{rotAzim}
\alias{rotAzim,RasterLayer-method}
\alias{rotAzim,PolarSegmentation-method}
\alias{rotAzim,BinImage-method}
\alias{rotAzim,AzimuthImage-method}
\title{Rotate Azimuth}
\usage{
rotAzim(x, cwRotAngle)

\S4method{rotAzim}{RasterLayer}(x, cwRotAngle)

\S4method{rotAzim}{PolarSegmentation}(x, cwRotAngle)

\S4method{rotAzim}{BinImage}(x, cwRotAngle)

\S4method{rotAzim}{AzimuthImage}(x, cwRotAngle)
}
\arguments{
\item{x}{\code{\linkS4class{RasterLayer}}.}
}
\value{
cwRotAngle \code{\linkS4class{Angle}}.
}
\description{
Rotate an \code{\linkS4class{AzimuthImage}} or a
similar \code{\linkS4class{RasterLayer}}.
}
\examples{
/inst/examples/rotAzimExample.R
}
