% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_read_cp.R
\docType{methods}
\name{slope}
\alias{slope}
\alias{slope<-}
\alias{slope,CanopyPhoto-method}
\alias{slope<-,CanopyPhoto,Angle-method}
\title{Set or get the slope slot.}
\usage{
slope(x)

\S4method{slope}{CanopyPhoto}(x)

\S4method{slope}{CanopyPhoto,Angle}(x) <- value
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{value}{\code{\linkS4class{Angle}} that represents the slope of
\code{x}.}
}
\value{
\code{\linkS4class{Angle}}.
}
\description{
Set or get the the slope of the photosite in which the
  photograph represented by a \code{\linkS4class{CanopyPhoto}} was taken.
}
\details{
The slope is the complement of the line-of-sight zenith angle.
  For zenith angle definition, see the details of \code{\link{lensPolyCoef}}.
}
\examples{
x <- loadPhoto()
elevation(x)
elevation(x) <- asAngle(55)
elevation(x)

\dontrun{
  elevation(x) <- asAngle(c(10,20))
}
}
\seealso{
\code{\link{loadPhoto}}.
}
