% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_per_pixel.R
\name{sRGB2HLS}
\alias{sRGB2HLS}
\alias{sRGB2HLS,matrix-method}
\alias{sRGB2HLS,numeric-method}
\alias{sRGB2HLS,CanopyPhoto-method}
\title{Convert sRGB to HLS.}
\usage{
sRGB2HLS(x, ...)

\S4method{sRGB2HLS}{matrix}(x)

\S4method{sRGB2HLS}{numeric}(x)

\S4method{sRGB2HLS}{CanopyPhoto}(x, ...)
}
\arguments{
\item{x}{numeric, matrix or \code{\linkS4class{CanopyPhoto}}. Values must
lying between \code{0} and \code{1}.}

\item{...}{Additional arguments as for \code{\link[raster]{writeRaster}}.}
}
\value{
Matrix, or \code{\linkS4class{CanopyPhoto}}.
}
\description{
Wrapper function of \code{\link[colorspace]{sRGB}} and
  \code{\link[colorspace]{HLS}} that convert colors from sRGB to HLS.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: Each row should represent a color. Argument \code{x} must have three columns.

\item \code{numeric}: Convert a single color from sRGB to HLS. Argument \code{x} must be of length three. The output is a matrix.

\item \code{CanopyPhoto}: The output is a \code{\linkS4class{CanopyPhoto}} with layer names H, L and S.
}}

\seealso{
\code{\link{normalize}}.

example /inst/examples/sRGB2HLSexample.R
}
