% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_per_pixel.R
\docType{methods}
\name{sRGB2LAB}
\alias{sRGB2LAB}
\alias{sRGB2LAB,matrix-method}
\alias{sRGB2LAB,numeric-method}
\alias{sRGB2LAB,CanopyPhoto-method}
\title{Convert sRGB to LAB.}
\usage{
sRGB2LAB(x, ...)

\S4method{sRGB2LAB}{matrix}(x)

\S4method{sRGB2LAB}{numeric}(x)

\S4method{sRGB2LAB}{CanopyPhoto}(x, ...)
}
\arguments{
\item{x}{numeric, matrix or \code{\linkS4class{CanopyPhoto}}. Values must
lying between \code{0} and \code{1}.}

\item{...}{Additional arguments as for \code{\link[raster]{writeRaster}}.}
}
\value{
Matrix, or \code{\linkS4class{CanopyPhoto}}.
}
\description{
Wrapper function of \code{\link[colorspace]{sRGB}} and
  \code{\link[colorspace]{LAB}} that convert colors from sRGB to LAB.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: Each row should represent a color. Argument \code{x} must have three columns.

\item \code{numeric}: Convert a single color from sRGB to LAB. Argument \code{x} must be of length three. The output is a matrix.

\item \code{CanopyPhoto}: The output is a \code{\linkS4class{CanopyPhoto}} with layer names L, A and B.
}}

\examples{
sRGB2LAB(c(0.5,0.7,0.5))
x <- sRGB2LAB(matrix(c(normalize(1:90, 1, 90)), ncol = 3))
str(x)
head(x)

x <- loadPhoto()
x <- normalize(x, 0, 255)
x <- sRGB2LAB(x)
x
plot(x)

x <- system.file("external/Black_caiman_Macrofotografie_2.jpg", package="caiman")
x <- loadPhoto(x)
plotRGB(x)
x <- normalize(x, 0, 255)
x <- sRGB2LAB(x)
plot(x)
}
\seealso{
\code{\link{normalize}}.
}
