% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAloadPhotoFunctions.R
\docType{methods}
\name{ssDenominator}
\alias{ssDenominator}
\alias{ssDenominator<-}
\alias{ssDenominator,CanopyPhoto-method}
\alias{ssDenominator<-,CanopyPhoto,numeric-method}
\title{Set or get the ssDenominator slot.}
\usage{
ssDenominator(x)

\S4method{ssDenominator}{CanopyPhoto}(x)

\S4method{ssDenominator}{CanopyPhoto,numeric}(x) <- value
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{value}{numeric. Denominator of the shutter speed used to take \code{x}.}
}
\value{
numeric.
}
\description{
Set or get the denominator of the shutter speed used to take the photograph represented by a \code{\linkS4class{CanopyPhoto}}.
}
\examples{
x <- loadPhoto()
ssDenominator(x)
ssDenominator(x) <- 125
ssDenominator(x)
}
\seealso{
\code{\link{loadPhoto}}.
}
