% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_MBLT.R
\docType{methods}
\name{subpixel2bin}
\alias{subpixel2bin}
\alias{subpixel2bin,RasterLayer-method}
\title{todo}
\usage{
subpixel2bin(subpixel, segmentation)

\S4method{subpixel2bin}{RasterLayer}(subpixel, segmentation)
}
\arguments{
\item{subpixel}{todo}

\item{segmentation}{todo}
}
\value{
\code{\linkS4class{BinImage}}
}
\description{
todo
}
\examples{
#todo
}
