% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_utilities.R
\name{switchUnit}
\alias{switchUnit}
\alias{switchUnit,Angle-method}
\title{Switch the unit of an Angle.}
\usage{
switchUnit(x)

\S4method{switchUnit}{Angle}(x)
}
\arguments{
\item{x}{\code{\linkS4class{Angle}}.}
}
\description{
Switch consistently the unit of an \code{\linkS4class{Angle}}.
  If it is in radians convert it to degrees and vice versa.
}
\examples{
a <- asAngle(90)
a
switchUnit(a)
}
\seealso{
\code{\link{asAngle}}.
}
