% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_classes.R
\docType{class}
\name{ZenithImage-class}
\alias{ZenithImage-class}
\title{An S4 class to represent zenith angles as an image.}
\description{
An S4 class to represent zenith angles as an image. Use the
  helper function \code{\link{makeZimage}} to generate new objects of this
  class.
}
\section{Slots}{

\describe{
\item{\code{lens}}{\code{\linkS4class{LensPolyCoef}}.}

\item{\code{...}}{Inherited from \code{\linkS4class{RelativeRadiusImage}}.}
}}

\examples{
showClass("ZenithImage")

}
\seealso{
\code{\link{makeZimage}}
}
