% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAperPixelProcessingFunctions.R
\docType{methods}
\name{adaptive_binarization}
\alias{adaptive_binarization}
\alias{adaptive_binarization,CanopyPhoto-method}
\title{Adaptive binarization}
\usage{
adaptive_binarization(cp, z, m, thr4only_catching_plants = 0.2)

\S4method{adaptive_binarization}{CanopyPhoto}(cp, z, m,
  thr4only_catching_plants = 0.2)
}
\arguments{
\item{cp}{\code{\linkS4class{CanopyPhoto}}. todo}

\item{z}{\code{\linkS4class{ZenithImage}}. todo}

\item{thr4only_catching_plants}{numerical.}

\item{mask}{\linkS4class{BinImage}. todo}
}
\value{
numeric or \code{\linkS4class{Raster}}.
}
\description{
todo
}
\details{
todo
}
\seealso{
\code{\link{autoThr}}, \code{\link{normalize}}.
}
