% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClasses.R
\docType{class}
\name{Angle-class}
\alias{Angle-class}
\title{An S4 class to represent angle values.}
\description{
An S4 class to represent angle values and made things easy for users because has a slot that set if it is in degrees or radians, so method and functions can adapt to it instead of users. Only allow positive values. Use the helper function \code{\link{asAngle}} to generate new objects of this class.
}
\section{Slots}{

\describe{
\item{\code{values}}{numeric.}

\item{\code{degrees}}{logical.}
}}
\examples{
showClass("Angle")

}
\seealso{
\code{\link{asAngle}}
}

