% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_read_cp.R
\name{bearing}
\alias{bearing}
\alias{bearing<-}
\alias{bearing,CanopyPhoto-method}
\alias{bearing<-,CanopyPhoto,Angle-method}
\title{Set or get the bearing slot.}
\usage{
bearing(x)

\S4method{bearing}{CanopyPhoto}(x)

\S4method{bearing}{CanopyPhoto,Angle}(x) <- value
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{value}{\code{\linkS4class{Angle}} that represents the bearing of
\code{x}.}
}
\value{
\code{\linkS4class{Angle}}.
}
\description{
Set or get the \code{\linkS4class{Angle}} that represents the
  bearing of a \code{\linkS4class{CanopyPhoto}}.
}
\details{
Bearing is a synonym for azimuth commonly used in geotagged
  photographs. Like azimuth, it is the angle between the North direction and
  the line of sight projected to the leveled plane. Since it is used for
  non-hemispherical-photographs, the line of sight refers to the centerline
  of the cone of vision (see details of \code{\link{lensPolyCoef}}). But if
  the photographs have been taken looking directly to the zenith (90 degrees
  of elevation), bearing does not make sense. In that case, bearing refers to
  the orientation of the top of the frame, which means taking photographs
  with the top of the camera facing the bearing. The same is valid for
  downward looking leveled hemispherical photographs.
}
\examples{
x <- loadPhoto()
bearing(x)
bearing(x) <- asAngle(10)
bearing(x)
\dontrun{
  bearing(x) <- asAngle(c(10,20))
}
}
\seealso{
\code{\link{makeAimage}}
}
