% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbres.R
\docType{methods}
\name{bgInterpol2}
\alias{bgInterpol2}
\alias{bgInterpol2,CanopyPhoto-method}
\title{Background interpolation}
\usage{
bgInterpol2(x, bin, mask, filling, prob = 0.95, erode = TRUE, fact = 1,
  np = 6)

\S4method{bgInterpol2}{CanopyPhoto}(x, bin, mask, filling, prob = 0.95,
  erode = TRUE, fact = 1, np = 6)
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{bin}{\code{\linkS4class{BinImage}}.}

\item{filling}{\code{\linkS4class{RasterLayer}} or \code{\linkS4class{RasterStack}}}

\item{prob}{numeric. Probability for \code[stats]{\link{quantile}}.}

\item{erode}{logical.}
}
\value{
\code{\linkS4class{BinImage}}
}
\description{
todo
}
\examples{
#todo
}
