% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAgeometryFunctions.R
\docType{methods}
\name{calcDiameter}
\alias{calcDiameter}
\alias{calcDiameter,LensPolyCoef-method}
\title{Calculate diameter corresponding to a circular fisheye photography.}
\usage{
calcDiameter(x, pix, angle)

\S4method{calcDiameter}{LensPolyCoef}(x, pix, angle)
}
\arguments{
\item{x}{\code{\linkS4class{LensPolyCoef}}.}

\item{pix}{todo}

\item{angle}{todo}
}
\value{
numeric.
}
\description{
Calculate diameter corresponding to a circular fisheye photography.
}
\examples{
/inst/examples/calcDiameterExample.R
}
