% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_utilities.R
\name{calcExposure}
\alias{calcExposure}
\alias{calcExposure,numeric-method}
\alias{calcExposure,CanopyPhoto-method}
\alias{calcExposure,character-method}
\title{Calculate exposure}
\usage{
calcExposure(x, exposureTime)

\S4method{calcExposure}{numeric}(x, exposureTime)

\S4method{calcExposure}{CanopyPhoto}(x)

\S4method{calcExposure}{character}(x)
}
\arguments{
\item{x}{numeric or character.}

\item{exposureTime}{numeric. The shutter speed.}
}
\value{
numeric
}
\description{
Calculate the exposure for a given aperture (fNumber) and
  exposure time (shutter speed).
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: \code{x} numeric. fNumbar, see
\code{\link{fNumber}}.

\item \code{CanopyPhoto}: Probably, this will be the most used.

\item \code{character}: \code{x} character. The names of the photographs
(extension included). If your working directory is the one that contains
the photographs, use just the file name, otherwise, use the full path to
the file.
}}

\examples{
x <- loadPhoto()
calcExposure(fNumber(x), exposureTime(x))
calcExposure(x)
}
\references{
Allen, E., & Triantaphillidou, S. (2010). The manual of
  photography. (E. Allen & S. Triantaphillidou, Eds.) (10th ed.). Amsterdam:
  Elsevier.
}
