% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_classes.R
\docType{class}
\name{CanopyPhoto-class}
\alias{CanopyPhoto-class}
\title{An S4 class to store vegetal canopy photographs.}
\description{
An S4 class to store vegetal canopy photographs. Use the
  helper function \code{\link{loadPhoto}} to build new objects of this class
  from a file.
}
\details{
\code{CanopyPhoto} was built on top of \code{RasterBrick} to take
  advantage of the \code{raster package} functionalities. However,
  \code{CanopyPhoto} requires raster implementation but not georeference and
  this is why some inherited slots are meaningless for this class.
}
\section{Slots}{

\describe{
\item{\code{equipment}}{character.}

\item{\code{fisheye}}{\code{\linkS4class{FishEye}}.}

\item{\code{datetime}}{character.}

\item{\code{geoLocation}}{\code{\link[sp]{SpatialPoints}}.}

\item{\code{bearing}}{\code{\linkS4class{Angle}}.}

\item{\code{elevation}}{\code{\linkS4class{Angle}}.}

\item{\code{slope}}{\code{\linkS4class{Angle}}.}

\item{\code{exposureTime}}{numeric.}

\item{\code{fNumber}}{numeric.}

\item{\code{isoSpeed}}{numeric.}

\item{\code{...}}{Inherited from \code{\linkS4class{RasterBrick}}.}
}}

\examples{
showClass("CanopyPhoto")

}
\seealso{
\code{\link{loadPhoto}}
}
