% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAloadPhotoFunctions.R
\docType{methods}
\name{datetime}
\alias{datetime}
\alias{datetime,CanopyPhoto-method}
\alias{datetime<-}
\alias{datetime<-,CanopyPhoto,character-method}
\title{Set or get datetime.}
\usage{
datetime(x)

\S4method{datetime}{CanopyPhoto}(x)

\S4method{datetime}{CanopyPhoto,character}(x) <- value
}
\arguments{
\item{x}{\code{\link{CanopyPhoto-class}}.}

\item{value}{Character that represents the elevation of \code{x}.}
}
\value{
Character.
}
\description{
Set or show the character string with date and time of shot for the photo represented by a \code{\linkS4class{CanopyPhoto}}.
}
\examples{
x <- loadPhoto()
datetime(x)
datetime(x) <- "1980-11-20 14:00:00"
datetime(x)

\dontrun{
  datetime(x) <- "an error"
}
}
\seealso{
\code{\link{loadPhoto}}.
}

