% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAutilityFunctions.R
\docType{methods}
\name{doMask}
\alias{doMask}
\alias{doMask,ZenithImage-method}
\title{It does a mask for hemispherical photographs.}
\usage{
doMask(z, a = makeAimage(z), zlim = NULL, alim = NULL)

\S4method{doMask}{ZenithImage}(z, a = makeAimage(z), zlim = NULL,
  alim = NULL)
}
\arguments{
\item{z}{\code{\linkS4class{ZenithImage}}.}

\item{a}{\code{\linkS4class{AzimuthImage}}.}

\item{zlim}{\code{\linkS4class{Angle}}. Set the zenith angle range with inclusive limits.}

\item{alim}{\code{\linkS4class{Angle}}. Set the azimuth angle range with inclusive limits.}
}
\value{
\code{\linkS4class{BinImage}}.
}
\description{
Use angular limits to do a mask for hemispherical photographs.
}
\details{
The intervals are closed on the right and on the left.
}
\examples{
x <- loadPhoto()
z <- makeZimage(ncol(x), lensPolyCoef(c(0.9192, -0.1792, -0.000443)))
m <- doMask(z, zlim = asAngle(c(30, 60)))
plot(m)

m <- doMask(z, zlim = asAngle(c(30, 60)), alim = asAngle(c(0, 90)))
plot(m)

plot(x[m, drop = FALSE])
}
\seealso{
\code{\link{makeZimage}}, \code{\link{makeAimage}}.
}

