% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAutilityFunctions.R
\docType{methods}
\name{doMask}
\alias{doMask}
\alias{doMask,ZenithImage-method}
\alias{doMask,CanopyPhoto-method}
\alias{doMask,RelativeRadiusImage-method}
\title{Do a mask for hemispherical photographs given angle restrictions.}
\usage{
doMask(x, z, previousMask = NULL, a = makeAimage(x), zlim = NULL,
  alim = NULL)

\S4method{doMask}{ZenithImage}(x, previousMask, a, zlim, alim)

\S4method{doMask}{CanopyPhoto}(x, z)

\S4method{doMask}{RelativeRadiusImage}(x)
}
\arguments{
\item{x}{\code{\linkS4class{ZenithImage}} or
\code{\linkS4class{CanopyPhoto}}.}

\item{previousMask}{\code{\linkS4class{BinImage}}. Default is \code{NULL}.}

\item{a}{\code{\linkS4class{AzimuthImage}}.}

\item{zlim}{\code{\linkS4class{Angle}}. Set the zenith angle range with
inclusive limits.}

\item{alim}{\code{\linkS4class{Angle}}. Set the azimuth angle range with
inclusive limits.}

\item{y}{todo}
}
\value{
\code{\linkS4class{BinImage}}.
}
\description{
Do a mask for hemispherical photographs given angle restrictions.
}
\details{
The intervals are closed on the right and on the left.

Todo: explain previousMask behavior.
}
\section{Methods (by class)}{
\itemize{
\item \code{ZenithImage}: You can use angular limits to do a mask for hemispherical
photographs.

\item \code{CanopyPhoto}: This feature is needed for further processing fullframe
hemispherical photographs.
}}

\examples{
x <- loadPhoto()
z <- makeZimage(ncol(x), lensPolyCoef(c(0.9192, -0.1792, -0.000443)))
m <- doMask(z, zlim = asAngle(c(30, 60)))
plot(m)

m <- doMask(z, zlim = asAngle(c(30, 60)), alim = asAngle(c(0, 90)))
plot(m)

plot(x[m, drop = FALSE])

# for fullframe hemispherical photos
path <- system.file("external/DSC_2881.jpg", package="caiman")
x <- loadPhoto(path)
### declaring it as a fullframe
fisheye(x) <- newFishEye(TRUE, TRUE, TRUE)

m <- doMask(x, z)
plot(m)

m <- doMask(z, previousMask = m, zlim = asAngle(c(30, 60)),
                                                    alim = asAngle(c(0, 90)))
plot(m)
}
\references{
Schneider, D., Schwalbe, E., Maas, H.-G., 2009. Validation of
  geometric models for fisheye lenses. ISPRS J. Photogramm. Remote Sens. 64,
  259-266.
}
\seealso{
\code{\link{makeZimage}}, \code{\link{makeAimage}}.
}
