% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doPolarQtree.R
\docType{methods}
\name{doPolarQtree}
\alias{doPolarQtree}
\alias{doPolarQtree,Raster-method}
\title{Quad-tree segmentation in a polar space.}
\usage{
doPolarQtree(x, z, a = makeAimage(z), scaleParameter, divisions = 1,
  mnSize = 1000)

\S4method{doPolarQtree}{Raster}(x, z, a = makeAimage(z), scaleParameter,
  divisions = 1, mnSize = 1000)
}
\arguments{
\item{x}{\linkS4class{Raster}. The raster that you want to process. It could have one or more layer.}

\item{z}{\code{\linkS4class{ZenithImage}}. This ZenithImage should match with the lens geometry of the picture linked with \code{x}.}

\item{a}{\code{\linkS4class{AzimuthImage}}.}

\item{scaleParameter}{one-length numeric. It forms part of the stopping condition (see Detalis).}

\item{divisions}{numeric. If it is 1, the segments could have a resolution of 30, 15, 7.5, 3.75 or 1.875 degrees. If it is 2, the segments could have a resolution of 15, 7.5, 3.75 or 1.875 degrees. If it is 3, the segments could have a resolution of 15, 7.5 or 3.75 degrees. If it is 4, the segments could have a resolution of 7.5 or 3.75 degrees.}

\item{mnSize}{numeric. This is an additional stoping criterion that avoids errors in segments near the zenith.}
}
\value{
\linkS4class{PolarSegmentation}.
}
\description{
The quad-tree segmentation algorithm is a top-down process that makes recursive divisions in four equal parts until a condition is satisfied and then stops locally. The usual implementation of the quad-tree algorithm is based on the raster structure, and this is why the resultant objects are squares of different sizes. This method implements the quad-tree segmentation but in a polar space.
}
\details{
Depending on \code{divisions}, the algorithm starts with segments of a given resolution. Next, it selects one segment and split it into four segments of equal angular resolution. Then, use the standard deviation of \code{x} as homogeneous criterion. To that end, calculate it for the entire segment and for each four segments. To stop the process locally, the algorithms evaluate if the sum of the standard deviation of the subsegments minus the standard deviation of the segment (\emph{delta}) is less or equal than the \code{scaleParameter}. If \code{x} is multilayer delta is calculated separately and the delta mean is used to evaluate the stopping condition.
}
\examples{
x <- loadPhoto()
z <- makeZimage(ncol(x), lensPolyCoef(c(0.6427, 0.0346, -0.024491)))
a <- makeAimage(z)
x <- normalize(x, 0, 255)
\donttest{
seg <- doPolarQtree(x, z, a, scaleParameter = 1)
seg
plot(seg)
}
# Next lines open a result obtained with scaleParameter = 0.1, which takes
# too long for an example.
seg <- raster(system.file("external/seg.tif", package="caiman"))
seg <- as(seg, "PolarSegmentation")
plot(seg)
}
\references{
Diaz, G.M., Lencinas, J.D., 2015. Enhanced Gap Fraction Extraction From Hemispherical Photography. IEEE Geosci. Remote Sens. Lett. 12, 1784-1789.
}
\seealso{
\code{\link{makeZimage}}, \code{\link{makeAimage}}.
}

