% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAperPixelProcessingFunctions.R
\docType{methods}
\name{enhanceHemiPhoto}
\alias{enhanceHemiPhoto}
\alias{enhanceHemiPhoto,CanopyPhoto-method}
\alias{enhanceHemiPhoto,matrix-method}
\title{Enhance upward looking hemispherical photographs.}
\usage{
enhanceHemiPhoto(x, ...)

\S4method{enhanceHemiPhoto}{matrix}(x, thr = NULL, fuzzyness = NULL,
  wR = 0.5, wB = 1.5, skyBlue = colorspace::sRGB(matrix(normalize(c(135,
  206, 235), 0, 255), ncol = 3)), ...)

\S4method{enhanceHemiPhoto}{CanopyPhoto}(x,
  mask = !is.na(makeRimage(ncol(x))), wR = 0.5, wB = 1.5,
  sharpen = TRUE, skyBlue = colorspace::sRGB(matrix(normalize(c(135, 206,
  235), 0, 255), ncol = 3)), ...)
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{...}{Additional arguments as for \code{\link[raster]{writeRaster}}.}

\item{thr}{numeric. By default, the algorithm will try to estimate it.}

\item{fuzzyness}{numeric. By default, the algorithm will try to estimate it.}

\item{wR}{numeric. Weight of red layer. See details.}

\item{wB}{numeric. Weight of blue layer. See details.}

\item{skyBlue}{\linkS4class{color}. See details.}

\item{mask}{\code{\linkS4class{BinImage}}.}

\item{sharpen}{logical. The default is TRUE, see details.}
}
\value{
Numeric or \linkS4class{RasterLayer}.
}
\description{
This algorithm uses the color perceptual attributes to enhance the contrast between the sky and plants through fuzzy classification. Color has three different perceptual attributes: hue, lightness, and chroma. The algorithm was developed using this premise: the color of the sky is different from the color of plants. It performs the next classification rules, here expressed in natural language: clear sky is blue and clouds decrease its chroma; if clouds are highly dense, then the sky is achromatic, and, in such cases, it can be light or dark; everything that does not match this description is not sky. These linguistic rules were translated to math language by means of fuzzy logic.
}
\details{
The greatest contrast on gray levels between the sky and the plants should be observed in the blue channel because sky scattering is relatively high in the higher frequencies of light, and so the blue channel ensures the brightest sky pixels. Nevertheless, in the blue channel, the interface between the sky and the foliage pixels can be as bright as sky pixels. These pixels look dark cyan colored instead of green. The color distortion observed should be linked with: color formation through color filter array and chromatic aberration. Thus, we propose to use an original feature that we denominate Relative Brightness instead of one-channel brightness, which use weighting parameters for red and blue channels (see references).

If sharpen is set to TRUE (default), a sharpen filter is applied to the raster with the membership values. The kernel used is \code{matrix(c(rep(-1, 3), -1, 12, -1, rep(-1, 3)), ncol = 3)}.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: The output is a numeric vector.

\item \code{CanopyPhoto}: The output is a \linkS4class{RasterLayer}.
}}
\examples{
x <- loadPhoto()
x <- normalize(x, 0, 255)
x <- enhanceHemiPhoto(x)
plot(x)
}
\references{
Diaz, G.M., Lencinas, J.D., 2015. Enhanced Gap Fraction Extraction From Hemispherical Photography. IEEE Geosci. Remote Sens. Lett. 12, 1784-1789.
}
\seealso{
\code{\link{membership2color}}, \code{\link{fuzzyLightness}}.
}

