% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAutilityFunctions.R
\docType{methods}
\name{expandFullframe}
\alias{expandFullframe}
\alias{expandFullframe,CanopyPhoto-method}
\title{Expand full frame hemispherical photographs}
\usage{
expandFullframe(x, zero = TRUE)

\S4method{expandFullframe}{CanopyPhoto}(x, zero = TRUE)
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{zero}{logical. By default is set as \code{TRUE} meaning that added
pixels will be \code{0}, if \code{FALSE} added pixels will be \code{NA}.}
}
\value{
\code{\linkS4class{CanopyPhoto}}.
}
\description{
Expand a full frame hemispherical photograph to get the
  equivalent of a circular hemispherical photograph.
}
\examples{
x <- loadPhoto()
z <- makeZimage(ncol(x), lensPolyCoef(c(0.9192, -0.1792, -0.000443)))
m <- doMask(z, zlim = asAngle(c(30, 60)))
plot(m)

m <- doMask(z, zlim = asAngle(c(30, 60)), alim = asAngle(c(0, 90)))
plot(m)

plot(x[m, drop = FALSE])

# Making an artificially fullframe hemiphoto by cropping a non-fullframe
# hemiphoto (just as an example)
path <- system.file("external/UnFavAutoE3.jpg", package="caiman")
x <- loadPhoto(path, upperLeft = c(33, 120), width = 414, height = 240)
## declaring it as a fullframe
fisheye(x) <- newFishEye(TRUE, TRUE, TRUE)

m <- doMask(x)
plot(m)

m <- doMask(z, previousMask = m, zlim = asAngle(c(30, 60)),
                                                    alim = asAngle(c(0, 90)))
plot(m)

## Diagonals of the rectangular frame are the radius of the circle:
plot(doMask(makeRimage(ncol(m))) + m, legend = FALSE)
}
\seealso{
\code{\link{doMask}}.
}
