% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_geometry.R
\name{expandFullframe}
\alias{expandFullframe}
\alias{expandFullframe,CanopyPhoto,ZenithImage-method}
\title{Expand full frame hemispherical photographs}
\usage{
expandFullframe(x, z, zenith_colrow)

\S4method{expandFullframe}{CanopyPhoto,ZenithImage}(x, z, zenith_colrow)
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{z}{\code{\linkS4class{ZenithImage}}.}
}
\value{
\code{\linkS4class{CanopyPhoto}}.
}
\description{
Expand a full frame hemispherical photograph to get the
  equivalent of a circular hemispherical photograph. Added
  pixels will be \code{0}.
}
\examples{
x <- loadPhoto()
z <- makeZimage(ncol(x), lensPolyCoef(c(0.9192, -0.1792, -0.000443)))
m <- doMask(z, zlim = asAngle(c(30, 60)))
plot(m)

m <- doMask(z, zlim = asAngle(c(30, 60)), alim = asAngle(c(0, 90)))
plot(m)

plot(x[m, drop = FALSE])

# for fullframe hemispherical photos
path <- system.file("external/DSC_2881.jpg", package="caiman")
x <- loadPhoto(path)
### declaring it as a fullframe
fisheye(x) <- newFishEye(TRUE, TRUE, TRUE)

m <- doMask(x, z)
plot(m)

m <- doMask(z, previousMask = m, zlim = asAngle(c(30, 60)),
                                                    alim = asAngle(c(0, 90)))
plot(m)
}
\seealso{
\code{\link{doMask}}.
}
