% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAobjectBasedProcessingFunctions.R
\docType{methods}
\name{extractFeatures}
\alias{extractFeatures}
\alias{extractFeatures,RasterLayer-method}
\title{Extract object features.}
\usage{
extractFeatures(x, ...)

\S4method{extractFeatures}{RasterLayer}(x, segmentation, fun)
}
\arguments{
\item{x}{\code{\linkS4class{RasterLayer}}. The image that you want extract feature from.}

\item{...}{Additional arguments.}

\item{segmentation}{\linkS4class{PolarSegmentation}. The segmentation that delimitate segments in \code{x}.  It must have the same extent and resolution of \code{x}.}

\item{fun}{function. It need to be able to take a single vector as input and return a one-length numeric or logical vector (e.g. mean).}
}
\value{
Numeric.
}
\description{
Given a single layer data, a segmentation and a function, this algorithm returns a numeric vector of length equal to the number of segment in the segmentation. Each value which is the return of processing all the pixels that belong to a segment with the aforementioned function.
}
\examples{
x <- loadPhoto()
z <- makeZimage(ncol(x), lensPolyCoef(c(0.9192, -0.1792, -0.000443)))
r <- makeRings(z, angleWidth = asAngle(30))
fi <- getFeatureImage(raster::subset(x, 3), r, sd)
plot(fi)
}
\seealso{
\code{\link{getFeatureImage}}.
}

