% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_read_cp.R
\docType{methods}
\name{fisheye}
\alias{fisheye}
\alias{fisheye<-}
\alias{fisheye,CanopyPhoto-method}
\alias{fisheye<-,CanopyPhoto,FishEye-method}
\title{Set or get the fisheye slot.}
\usage{
fisheye(x)

\S4method{fisheye}{CanopyPhoto}(x)

\S4method{fisheye}{CanopyPhoto,FishEye}(x) <- value
}
\arguments{
\item{x}{\code{\linkS4class{CanopyPhoto}}.}

\item{value}{\code{\linkS4class{FishEye}}.}
}
\value{
\code{\linkS4class{FishEye}}.
}
\description{
Set or get the fisheye object that describe the photograph
  represented by a \code{\linkS4class{CanopyPhoto}}.
}
\examples{
x <- loadPhoto()
fisheye(x)
fisheye(x) <- newFishEye(TRUE, FALSE, FALSE)
fisheye(x)
}
\seealso{
\code{\link{loadPhoto}}.
}
