% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_utilities.R
\name{getMin}
\alias{getMin}
\alias{getMax}
\alias{getMin,RasterAll-method}
\alias{getMax,RasterAll-method}
\title{Get the extreme values.}
\usage{
getMin(x)

\S4method{getMin}{RasterAll}(x)

getMax(x)

\S4method{getMax}{RasterAll}(x)
}
\arguments{
\item{x}{\code{\linkS4class{Raster}}.}
}
\value{
A numeric vector.
}
\description{
Get the maximum/minimum value of a \code{\linkS4class{Raster}}.
  If it is multilayer, then get the maximums/minimums values of each layer.
}
\examples{
x <- loadPhoto()
getMax(x)
getMin(x)

x <- makeRimage(480)
getMax(x)
getMin(x)
}
