% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_segmentation.R
\name{makePolarSectors}
\alias{makePolarSectors}
\alias{makePolarSectors,AzimuthImage-method}
\title{Slice the polar space in sectors}
\usage{
makePolarSectors(x, angleWidth, angleMean = FALSE)

\S4method{makePolarSectors}{AzimuthImage}(x, angleWidth, angleMean = FALSE)
}
\arguments{
\item{x}{\code{\linkS4class{AzimuthImage}}.}

\item{angleWidth}{\code{\linkS4class{Angle}}. It must divides \code{0:360}
into a whole number of segments.}

\item{angleMean}{logical. If \code{FALSE}, all the pixels that belong to a sector
are labeled with an ID number. Otherwise, the angle mean of the sector are
assigned to the pixels.}
}
\value{
\code{\linkS4class{PolarSegmentation}}.
}
\description{
Make sectors by slicing the azimuth angle from \code{0} to
  \code{360} in equals intervals.
}
\details{
The intervals are closed on the right and open on the left. The
  first sector never contains \code{0} because the zenith point is always in
  between two pixels (see \code{\link{makeZimage}}).
}
\examples{
z <- makeZimage(480, lensPolyCoef())
a <- makeAimage(z)

s <- makePolarSectors(a, angleWidth = asAngle(10))
s
plot(s)
s <- makePolarSectors(a, angleWidth = asAngle(10), angleMean = TRUE)
s
plot(s)

s <- makePolarSectors(a, angleWidth = asAngle(15))
s
plot(s)
}
\seealso{
\code{\link{makeAimage}}, \code{\link{makeRings}}, \code{\link{makePolarGrid}}.
}
