% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AApolarSegmentationAlgorithms.R
\docType{methods}
\name{makeRings}
\alias{makeRings}
\alias{makeRings,ZenithImage-method}
\title{Make rings.}
\usage{
makeRings(x, angleWidth, angleMean = FALSE)

\S4method{makeRings}{ZenithImage}(x, angleWidth, angleMean = FALSE)
}
\arguments{
\item{x}{\code{\linkS4class{ZenithImage}}.}

\item{angleWidth}{\code{\linkS4class{Angle}}. AngleWidth must divide 0-90 into a whole number of segments.}

\item{angleMean}{logical. If it is FALSE, all the pixels belong to a ring are labeled with an ID number. Otherwise, the angle mean of the ring are assigned to the pixels.}
}
\value{
\code{\linkS4class{PolarSegmentation}}.
}
\description{
Make rings by dividing the zenith angle from 0 to 90 in equals intervals.
}
\details{
The intervals are closed on the right and open on the left. The first ring never contain \code{0} because zenith point is always in between two pixels (\code{\link{makeZimage}}).
}
\examples{
z <- makeZimage(480, lensPolyCoef())

r <- makeRings(z, angleWidth = asAngle(30))
r
plot(r)
r <- makeRings(z, angleWidth = asAngle(30), angleMean = TRUE)
r
plot(r)

r <- makeRings(z, angleWidth = asAngle(10))
plot(r)
}
\seealso{
\code{\link{makeZimage}}, \code{\link{makePolarSectors}}, \code{\link{makePolarGrid}}.
}

