% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_subgroup_utilities.R
\name{rtitle}
\alias{rtitle}
\alias{rtitle<-}
\alias{rtitle,Raster-method}
\alias{rtitle<-,Raster,character-method}
\title{Set or get the title of a Raster*.}
\usage{
rtitle(x)

\S4method{rtitle}{Raster}(x)

\S4method{rtitle}{Raster,character}(x) <- value
}
\arguments{
\item{x}{\code{\linkS4class{Raster}}.}

\item{value}{character. The title.}
}
\value{
character.
}
\description{
Set or get the title of a \code{\linkS4class{Raster}}.
}
\details{
This method was written after reading this
  \href{https://stackoverflow.com/questions/9900134/is-it-bad-practice-to-access-s4-objects-slots-directly-using}{stackoverflow
  answer}.
}
\examples{
x <- makeRimage(480)
rtitle(x)
rtitle(x) <- "some string"
rtitle(x)
}
